refactor: 카테고리 조회 로직을 모의 데이터에서 실제 API 호출로 변경
- fetchInstance를 사용하여 실제 API 호출 구현
- getCategories 함수를 실제 API 엔드포인트를 호출하도록 수정
- useGetCategories 훅의 이름을 useCategoriesQuery로 변경하여 네이밍 규칙 적용
- 모의 데이터 관련 코드 제거
- 쿼리 키 간소화
